Use built-in variables and common tools, translate HOSTTYPE
HOSTTYPE (and OSTYPE, but that's superseded by OSD_FAMILY) is a
built-in variable we use when available, which needs some translation
to Go's idioms.

Further, calling whoami is not necessary as there's built-ins.

To tackle the case of a cache getting restored from the wrong arch
we check whether "go version" runs at all.

'go' (the binary) might've come from a different directory than
we use, there it's less error-prone to gate deletion differently,
by its mere existence.